Iterate over ALL wireguard peers
Before this commit, only the first chunk of wireguard peers was
processed. The following chunks of peers were silently discarded.

Currently we see chunks that approximately have a size of 90 peers.
So you will only observe this issue if you have more than 90 peers.

note: this is a modified version in order to match wgkex code

Co-authored-by: "aiyion.prime" <git@aiyionpri.me>