worker: use get_attr and prevent TypeError
The `WG<XXX>` attributes are no longer present after
https://github.com/svinota/pyroute2/commit/3e485542722e3cdc58b1277b12366424a176fe23
so .get_attr has to beed used in order to remain compatible with
pyroute2 >= 0.5.15.

Calling extend on a list with None as argument waill raise a TypeError.
This happens when no peers are added yet (e.g. after a fresh boot)